Add files via upload
The project is related to create a Game called Guessing The Random Number which is generated by the Computer System. The generated random number is between 0 to 100. If you guess the number you will get points on the basis of your number of attemps. I have created a function and call it as per requirement in the code. You have maximum of 10 attemps to guess the random number.